Tweaks

+++ src/hotspot/cpu/aarch64/assembler_aarch64.hpp
@@ -926,7 +926,7 @@ public:
   static const uint64_t branch_range = NOT_DEBUG(128 * M) DEBUG_ONLY(2 * M);
 
   static bool reachable_from_branch_at(address branch, address target) {
-    return uabs(target - branch) < branch_range;
+    return uabs((uint64_t)(target - branch)) < branch_range;
   }
 
   // Unconditional branch (immediate)
